requirements.txt
Django==2.2.6
pytz==2019.3
sqlparse==0.3.0
typing==3.7.4.1
djangorestframework==3.10.3
protobuf==3.10.0
protobuf3-to-dict==0.1.5
protoc-gen-swagger==0.1.0
retrying==1.3.3
python-dateutil==2.8.0
psycopg2-binary==2.8.3
grpclib==0.3.0
grpcio==1.24.1
grpcio-tools==1.24.1
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
googleapis-common-protos==1.6.0
